Test: Syntax (see desc)
About:
A syntax test is when ShopSearch is put into a testing server and
checked for errors such as indentations, speech marks and brackets.

Errors:
Starting - 22
Ending - 0
Fix rate: 100%

NOTE: This does not mean that ShopSearch_v2.0 is safe to use as a Logic
test hasn't been ran on it.